feat: 게시글 유효성 검증 로직 개선

+++ src/main/java/cotw/server/domain/board/dto/response/PostListResponseDto.java
@@ -18,10 +18,10 @@ public class PostListResponseDto {
     private final LocalDateTime createdAt;
 
     // 금액 관련
-    private final int amount;        // 목표 금액
-    private final int currentAmount;        // 현재 모금액
-    private final int remaining;     // 남은 금액(음수면 0)
-    private final int overfunded;    // 초과 모금액(<=0이면 0)
+    private final long  amount;        // 목표 금액
+    private final long  currentAmount;        // 현재 모금액
+    private final long  remaining;     // 남은 금액(음수면 0)
+    private final long  overfunded;    // 초과 모금액(<=0이면 0)
     private final double percent;    // 게이지용 0~100, 소수점 1자리
     private final double percentRaw; // 실제 달성률(100 초과 가능)
 
@@ -46,7 +46,7 @@ public class PostListResponseDto {
         this.createdAt = post.getCreatedAt();
         this.donorCount = post.getDonorCount();
         // 남은 금액(음수 방지)
-        int rawRemaining = amount - currentAmount;
+        long  rawRemaining = amount - currentAmount;
         this.remaining = Math.max(rawRemaining, 0);
 
         // 실제 달성률과 게이지용 퍼센트
