+++ src/main/java/cotw/server/domain/board/dto/response/PostResponseDto.java
@@ -38,10 +38,10 @@ public class PostResponseDto {
     private LocalDateTime updatedAt;
 
     // 금액 관련 (계산된 값들)
-    private int amount;           // 목표 금액
-    private int currentAmount;    // 현재 모금액
-    private int remaining;        // 남은 금액 (음수면 0)
-    private int overfunded;       // 초과 모금액 (<=0이면 0)
+    private long amount;           // 목표 금액
+    private long currentAmount;    // 현재 모금액
+    private long remaining;        // 남은 금액 (음수면 0)
+    private long overfunded;       // 초과 모금액 (<=0이면 0)
     private double percent;       // 게이지용 퍼센트 (0~100, 소수점 1자리)
     private double percentRaw;    // 실제 달성률 (100 초과 가능)
 
@@ -86,7 +86,7 @@ public class PostResponseDto {
         this.currentAmount = post.getCurrentAmount();
 
         // 남은 금액 (음수 방지)
-        int rawRemaining = amount - currentAmount;
+        long  rawRemaining = amount - currentAmount;
         this.remaining = Math.max(rawRemaining, 0);
 
         // 달성률 계산
